sagemaker/notebook/requirements.txt
aws-cdk.aws-ec2==1.130.0
aws-cdk.aws-iam==1.130.0
aws-cdk.aws-s3==1.130.0
aws-cdk.aws-sagemaker==1.130.0
aws-cdk.core==1.130.0
